Compile fixes
