+++ Sources/TripKit/categories/Optional+OrThrow.swift
