+++ Sources/TripKit/categories/String+NonEmpty.swift
