+++ Sources/TripKitAPI/TKRegionManager.swift
@@ -63,6 +63,10 @@ public class TKRegionManager: NSObject {
     return response?.regions ?? []
   }
   
+  public func modeDetails(forModeIdentifier mode: String) -> TKAPI.ModeDetails? {
+    return response?.modes?[mode]
+  }
+  
   public var regionsHash: NSNumber? {
     if let hashCode = response?.hashCode {
       return NSNumber(value: hashCode + 2) // Force update after broken polygons
@@ -179,32 +183,17 @@ extension TKRegionManager {
 // MARK: - Testing coordinates
 
 extension TKRegionManager {
- 
-#if canImport(CoreLocation)
-  public func coordinateIsPartOfAnyRegion(_ coordinate: CLLocationCoordinate2D) -> Bool {
-    for region in regions {
-      if region.contains(coordinate) {
-        return true
-      }
-    }
-    return false
-  }
-#endif
   
-#if canImport(MapKit)
-  /// Used to check if user can route in that area.
-  public func mapRectIntersectsAnyRegion(_ mapRect: MKMapRect) -> Bool {
-    // TODO: How to handle rect spanning 180th medidian?
+  public func coordinateIsPartOfAnyRegion(latitude: TKAPI.Degrees, longitude: TKAPI.Degrees) -> Bool {
     for region in regions {
-      if region.intersects(mapRect) {
+      if region.contains(latitude: latitude, longitude: longitude) {
         return true
       }
     }
     return false
   }
-#endif
 }
-
+  
 // MARK: - Getting regions by coordinates, etc.
 
 extension TKRegionManager {
@@ -221,100 +210,5 @@ extension TKRegionManager {
   public func localRegion(code: String) -> TKRegion? {
     return regions.first { $0.code == code }
   }
-  
-#if canImport(MapKit)
-  /// - Returns: A matching local region or the shared instance of `TKInternationalRegion` if no local region contains this coordinate region.
-  public func region(containing region: MKCoordinateRegion) -> TKRegion {
-    return self.region(containing: region.topLeft, region.bottomRight)
-  }
-
-  /// - Returns: Local regions that overlap with the provided coordinate region. Can be empty.
-  public func localRegions(overlapping region: MKCoordinateRegion) -> [TKRegion] {
-    let mapRect = MKMapRect.forCoordinateRegion(region)
-    return regions.filter { $0.intersects(mapRect) }
-  }
-#endif
-
-#if canImport(CoreLocation)
-  /// Determines the local (non-international) regions for the coordinate pair
-  ///
-  /// - Parameters:
-  ///   - start: A valid coordinate
-  ///   - end: Another valid coordinate
-  /// - Returns: An array of either A) no element (if both coordinates are in the
-  ///     international region), B) one element (if both coordinates are in the
-  ///     the same region, or C) two elements (a local region for the start and
-  ///     one for the end coordinates).
-  public func localRegions(start: CLLocationCoordinate2D, end: CLLocationCoordinate2D) -> [TKRegion] {
     
-    let startRegions  = localRegions(containing: start)
-    let endRegions    = localRegions(containing: end)
-    
-    if let intersectingRegion = startRegions.intersection(endRegions).first {
-      return [intersectingRegion]
-    
-    } else {
-      return [startRegions, endRegions].compactMap { $0.first }
-    }
-    
-  }
-
-  /// - Parameter coordinate: A coordinate
-  /// - Returns: The local (non-international) regions intersecting with the 
-  ///     provided coordinate
-  public func localRegions(containing coordinate: CLLocationCoordinate2D) -> Set<TKRegion> {
-    guard coordinate.isValid else { return [] }
-    let containing = regions.filter { $0.contains(coordinate) }
-    return Set(containing)
-  }
-  
-  /// Determines a region (local or international) for the coordinate pair
-  ///
-  /// - Parameters:
-  ///   - first: A valid coordinate
-  ///   - second: Another valid coordinate
-  /// - Returns: A local region if both lie within the same or the shared
-  ///     international region instance.
-  public func region(containing first: CLLocationCoordinate2D, _ second: CLLocationCoordinate2D) -> TKRegion {
-    let local = localRegions(start: first, end: second)
-    if local.count == 1 {
-      return local.first!
-    } else {
-      return .international
-    }
-  }
-  
-  
-  /// - Parameter coordinate: A valid coordinate
-  /// - Returns: The time zone of a matching region for this coordinate. Will 
-  ///     return `nil` if the coordinate falls outside any supported region.
-  public func timeZone(for coordinate: CLLocationCoordinate2D) -> TimeZone? {
-    return regions.first { $0.contains(coordinate) }?.timeZone
-  }
-  
-  
-  /// Find city closest to provided coordinate, in same region
-  ///
-  /// - Parameter target: Coordinate for which to find closest city
-  /// - Returns: Nearest City
-  public func city(nearestTo target: CLLocationCoordinate2D) -> TKRegion.City? {
-    typealias Match = (TKRegion.City, CLLocationDistance)
-    
-    let cities = localRegions(containing: target).reduce(into: [TKRegion.City]()) { cities, region in
-      cities.append(contentsOf: region.cities)
-    }
-    let best = cities.reduce(nil) { acc, city -> Match? in
-      guard let distance = target.distance(from: city.coordinate) else { return acc }
-      
-      if let existing = acc?.1, existing < distance {
-        return acc
-      } else {
-        return (city, distance)
-      }
-    }
-    
-    return best?.0
-  }
-#endif    
-  
 }
